% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAppearance.R
\name{checkAppearance}
\alias{checkAppearance}
\title{checkAppearance}
\usage{
checkAppearance(x)
}
\arguments{
\item{x}{A code list as returned by \code{\link{codeExtract}}}
}
\value{
A list with four elements: appearance, setappearance, type and
warnings. Appearance is a matrix containing values which indicate whether an 
object appears in a part of the code or not (e.g. indicates whether "vm_example" 
appears in realization "on" of module "test" or not.). 0 means that it does not appear,
1 means that it appears in the code and 2 means that it appears in the
not_used.txt. setappearance contains the same information but for sets instead of other
objects. Type is a vector containing the type of each object (exluding sets). And warnings
contains a list of warnings created during that process.
}
\description{
Checks for all declared objects in which parts of the model they appear and
calculates the type of each object (core object, interface object, module
object of module xy,...)
}
\seealso{
\code{\link{codeCheck}},\code{\link{readDeclarations}}
}
\author{
Jan Philipp Dietrich
}
