% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDescription.R
\name{checkDescription}
\alias{checkDescription}
\title{checkDescription}
\usage{
checkDescription(x, w = NULL)
}
\arguments{
\item{x}{GAMS declarations matrix as returned by}

\item{w}{a vector of warnings the warnings should be added to
\code{\link{readDeclarations}}}
}
\value{
vector of warnings
}
\description{
Checks whether all Declarations of a GAMS code come with a Description,
throws out a warning in case of a missing description.
}
\seealso{
\code{\link{codeCheck}}
}
\author{
Jan Philipp Dietrich
}
