% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseFromList.R
\name{chooseFromList}
\alias{chooseFromList}
\title{chooseFromList}
\usage{
chooseFromList(
  theList,
  type = "runs",
  returnBoolean = FALSE,
  multiple = TRUE,
  allowEmpty = FALSE,
  group = FALSE,
  userinput = FALSE
)
}
\arguments{
\item{theList}{list to be selected from}

\item{type}{string to be shown to user to understand what they choose}

\item{returnBoolean}{TRUE: returns list with dimension of theList with FALSE or TRUE
FALSE: returns selected entries of theList}

\item{multiple}{TRUE: allows to select multiple entries. FALSE: no}

\item{allowEmpty}{TRUE: allows you not to select anything (returns NA). FALSE: must select something}

\item{group}{list with same dimension as theList with group names to allow to select whole groups, optional}

\item{userinput}{string provided by the user. If not supplied, user is asked}
}
\value{
either a boolean list with same length as input or only the selected list items.
}
\description{
Allows the user to select single or multiple elements from a list.
Entries can be selected based on individual choice, groups, regex or all.
}
\author{
Oliver Richters
}
