% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeExtract.R
\name{codeExtract}
\alias{codeExtract}
\title{codeExtract}
\usage{
codeExtract(codeFiles, name)
}
\arguments{
\item{codeFiles}{A vector of file names of GAMS code files.}

\item{name}{A name indicating what collection of code files this is (e.g.
module name)}
}
\value{
A list with two elements: code and declarations. Code contains the
cleaned up gams code and declarations contains the declarations matrix as
returned by \code{\link{readDeclarations}}
}
\description{
Returns aggregated and cleaned GAMS code together with declaration matrix
}
\seealso{
\code{\link{codeCheck}},\code{\link{readDeclarations}}
}
\author{
Jan Philipp Dietrich
}
