% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.modularGAMS.R
\name{convert.modularGAMS}
\alias{convert.modularGAMS}
\title{convert.modularGAMS}
\usage{
convert.modularGAMS(path = ".", modulepath = "modules/")
}
\arguments{
\item{path}{path to the main folder of the model}

\item{modulepath}{Module path within the model (relative to the model main
folder)}
}
\description{
Converts modular GAMS code from an older modular definition to the newest one
}
\seealso{
\code{\link{codeCheck}}
}
\author{
Jan Philipp Dietrich
}
