% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_input.R
\name{copy_input}
\alias{copy_input}
\title{copy_input}
\usage{
copy_input(x, sourcepath, suffix = NULL, move = FALSE)
}
\arguments{
\item{x}{Filepath or data frame containing the mapping of files to be deleted}

\item{sourcepath}{Path to folder containing all input files}

\item{suffix}{suffix that might be part of input names that should be deleted}

\item{move}{If TRUE files will be moved instead of copied (default=FALSE)}
}
\description{
Function to copy input files to their destination folders
}
\author{
Jan Philipp Dietrich, David Klein
}
