% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_olddata.R
\name{delete_olddata}
\alias{delete_olddata}
\title{delete_olddata}
\usage{
delete_olddata(x)
}
\arguments{
\item{x}{Filepath or data frame containing the mapping of files to be deleted}
}
\description{
Delete data provided in mapping
}
\author{
Jan Philipp Dietrich, David Klein
}
