% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAMScodeFilter.R
\name{GAMScodeFilter}
\alias{GAMScodeFilter}
\title{GAMScodeFilter}
\usage{
GAMScodeFilter(x)
}
\arguments{
\item{x}{A vector with lines of GAMS code (as you get by reading the code
with readLines)}
}
\value{
The cleaned GAMS code
}
\description{
Cleans GAMS code supplied from empty lines and comments.
}
\examples{

GAMScodeFilter(c("","*comment","a=12;","","b=13;"))

}
\seealso{
\code{\link{readDeclarations}}
}
\author{
Jan Philipp Dietrich
}
