% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GB_NPMLE_Load.R
\name{GB_NPMLE_Load}
\alias{GB_NPMLE_Load}
\title{Loading function of Generative Bootstrap NPMLE}
\usage{
GB_NPMLE_Load(save_filePy=path_py, save_fileR=path_r, option="Train", gpu_ind=0,
N=100, M=100, num_it=2000, verb=1, tol=0.005, lrDecay=0, boot_size=NULL)
}
\arguments{
\item{save_filePy}{the saved .pt/ckpt file directory from GB-NPMLE function.}

\item{save_fileR}{the saved .RData file directory from GB-NPMLE function.}

\item{option}{option = "Train": load a pre-trained generator and further train; option = "Sample": load a pre-trained generator and generate bootstrap samples.}

\item{gpu_ind}{gpu index.}

\item{N}{number of Monte Carlo iterations for approximating E_w.}

\item{M}{number of Monte Carlo iterations for approximating E_z.}

\item{num_it}{number of iterations for training.}

\item{verb}{print information while training generator.}

\item{tol}{tolerance to determine whether EM algorithm converges;default is 0.005.}

\item{lrDecay}{lrDecay = 1: using decaying learning rate.}

\item{lrpower}{decay rate of learning rate, default is 0.2.}

\item{boot_size}{the number of bootstrap samples to be generated.}
}
\description{
Loading a pre-trained generator for further training or Sample
}
\examples{
### Pre-training stage
library(reticulate)
set.seed(2^2+2021)
sigma = 0.5;n = 100
theta = c(rep(0,0.2*n),
          rep(5,0.8*n))
Y = theta+rnorm(n,0,sigma)
fit_GBnpmle = GB_NPMLE(Y=Y, param=sigma, distribution="Gaussian location",
num_it=200, n=n, p=1, S=100, q=100, verb=1, hidden_size=500,
save=1, save_path=save_path)

### Loading stage
fit_load = GB_NPMLE_Load(save_filePy=path_py, save_fileR=path_r, option="Train",
N=100, M=100, num_it=1800, verb=1, tol=0.005, lrDecay=0, boot_size=1000, gpu_ind=0)

Sample_GBnpmle = GB_NPMLE_Sample(fit_load[[1]], boot_size=10000)
hist(Sample_GBnpmle$Theta, breaks=25, main="", xlab=expression(theta), freq=F, col="white", border=T)
#hist(fit_load$Theta, breaks=25, main="", xlab=expression(theta), freq=F, col="white", border=T)
points(x=0, y=0, pch=4, col='red', lwd=2)
points(x=5, y=0, pch=4, col='red', lwd=2)

}
\seealso{
\code{\link{GB_NPMLE}}, \code{\link{GB_NPMLE_Sample}}
}
\author{
Shijie Wang and Minsuk Shin
}
