% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GB_NPMLE_Sample.R
\name{GB_NPMLE_Sample}
\alias{GB_NPMLE_Sample}
\title{Sampling function of Generative Bootstrap NPMLE}
\usage{
GB_NPMLE_Sample(fit_GBnpmle, boot_size=10000)
}
\arguments{
\item{fit_GBnpmle}{a GB_NPMLE object evaluated by "GB_NPMLE" function.}

\item{boot_size}{the number of bootstrap samples to be generated.}
}
\description{
Generate bootstrap samples from a trained NPMLE generator
}
\examples{
#sample_GBnpmle = GB_NPMLE_Sample(fit_GBnpmle, boot_size=10000)
}
\seealso{
\code{\link{GB_NPMLE}}, \code{\link{GB_NPMLE_Load}}
}
\author{
Shijie Wang and Minsuk Shin
}
