% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{generator}
\alias{generator}
\title{Generator of GBS for a given weight}
\usage{
generator(fit_GBS, w, lam, eta = NULL)
}
\arguments{
\item{fit_GMS}{the trained GMS object.}

\item{w}{weights: a B X S matrix.}

\item{lam}{tuning parameter: a scalar.}

\item{eta}{auxiliary parameter: a scalar.}
}
\description{
Generate the weighted M-estimator for a given weight and a tuning parameter.
}
\examples{
theta_hat = generator(fit_GMS, w=matrix(1,1,S), verb=0)
}
\seealso{
\code{\link{GMS_Sampling}}, \code{\link{GMS}}, \code{\link{post_process}}, \code{\link{GMS_Loading}}
}
\author{
Minsuk Shin, Jun Liu and Shijie Wang
}
