% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_info}
\alias{get_info}
\title{get_info}
\usage{
get_info(file, grep_expression, sep, pattern = "", replacement = "")
}
\arguments{
\item{file}{path to info.txt (including info.txt)}

\item{grep_expression}{String before the information that should be extracted}

\item{sep}{Separator between grep_expression and information}

\item{pattern}{Optional pattern to be replaced (default pattern = "")}

\item{replacement}{Optional replacement (default replacement = "")}
}
\description{
Function to extract information from info.txt
}
\author{
Jan Philipp Dietrich, David Klein
}
