% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfiledestinations.R
\name{getfiledestinations}
\alias{getfiledestinations}
\title{getfiledestinations}
\usage{
getfiledestinations()
}
\description{
Create file2destination mapping based on information from the model, ignoring
top-level directories listed in \code{.gitignore}.
}
\author{
Jan Philipp Dietrich, David Klein
}
