% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLine.R
\name{getLine}
\alias{getLine}
\title{getLine}
\usage{
getLine()
}
\value{
The user input as a string.
}
\description{
Get one line of user input regardless of whether running interactively or not (via Rscript).
(base::readline does not wait for user input when running via Rscript.)
}
