% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModules.R
\name{getModules}
\alias{getModules}
\title{getModules}
\usage{
getModules(modulepath)
}
\arguments{
\item{modulepath}{The path where the modules are stored.}
}
\value{
A matrix containing the different modules with name, corresponding
module number and corresponding realizations
}
\description{
Extract module information of a GAMS model.
}
\seealso{
\code{\link{codeCheck}}
}
\author{
Jan Philipp Dietrich
}
