% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gms-package.R
\docType{package}
\name{gms-package}
\alias{gms}
\alias{gms-package}
\title{gms: 'GAMS' Modularization Support Package}
\description{
A collection of tools to create, use and maintain modularized model code written in the modeling language 'GAMS' (\url{https://www.gams.com/}). Out-of-the-box 'GAMS' does not come with support for modularized model code. This package provides the tools necessary to convert a standard 'GAMS' model to a modularized one by introducing a modularized code structure together with a naming convention which emulates local environments. In addition, this package provides tools to monitor the compliance of the model code with modular coding guidelines.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/gms}
  \item \doi{10.5281/zenodo.4390032}
  \item Report bugs at \url{https://github.com/pik-piam/gms/issues}
}

}
\author{
\strong{Maintainer}: Jan Philipp Dietrich \email{dietrich@pik-potsdam.de} (\href{https://orcid.org/0000-0002-4309-6431}{ORCID}) (affiliation: Potsdam Institute for Climate Impact Research)

Authors:
\itemize{
  \item Jan Philipp Dietrich \email{dietrich@pik-potsdam.de} (\href{https://orcid.org/0000-0002-4309-6431}{ORCID}) (affiliation: Potsdam Institute for Climate Impact Research)
  \item David Klein (affiliation: Potsdam Institute for Climate Impact Research)
  \item Anastasis Giannousakis
  \item Felicitas Beier (affiliation: Potsdam Institute for Climate Impact Research)
  \item Johannes Koch (affiliation: Potsdam Institute for Climate Impact Research)
  \item Lavinia Baumstark (affiliation: Potsdam Institute for Climate Impact Research)
  \item Mika Pflüger
  \item Oliver Richters (affiliation: Potsdam Institute for Climate Impact Research)
}

}
