% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.modularGAMS.R
\name{is.modularGAMS}
\alias{is.modularGAMS}
\title{is.modularGAMS}
\usage{
is.modularGAMS(path = ".", version = FALSE, modulepath = "modules/")
}
\arguments{
\item{path}{path to the main folder of the model}

\item{version}{if TRUE returns the version of the modular structure or FALSE, 
otherwise returns a boolean indicating whether it is modular or not.}

\item{modulepath}{Module path within the model (relative to the model main
folder)}
}
\description{
Checks whether a folder seems to contain modular GAMS code or not.
}
\examples{
is.modularGAMS(system.file("dummymodel",package="gms"))

}
\seealso{
\code{\link{codeCheck}}
}
\author{
Jan Philipp Dietrich
}
