% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_dependencies.R
\name{model_dependencies}
\alias{model_dependencies}
\title{Function to detect R package dependencies}
\usage{
model_dependencies(mainfolder = ".")
}
\arguments{
\item{mainfolder}{main folder of the model to be analyzed}
}
\value{
A list of dependencies sorted by appearances
}
\description{
This function analyzes a model folder and all subfolders and searches
for library and require statements.
}
\author{
Jan Philipp Dietrich
}
