% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_interfaceplot.R
\name{modules_interfaceplot}
\alias{modules_interfaceplot}
\title{modules_interfaceplot}
\usage{
modules_interfaceplot(
  x = ".",
  modulepath = "modules",
  filetype = "png",
  targetfolder = NULL,
  writetable = TRUE,
  includeCore = FALSE,
  ...
)
}
\arguments{
\item{x}{Either the object returned by \code{\link{codeCheck}} or the path to
the main folder of the model.}

\item{modulepath}{Path to the modules folder}

\item{filetype}{Filetype that should be used (e.g. "png" or "pdf")}

\item{targetfolder}{Folder outputs should be written to. If set to NULL
outputs will be added to corresponding module folders}

\item{writetable}{Logical deciding whether a csv containing the interfaces
should be written as well.}

\item{includeCore}{Logical to create plot for core or not, default FALSE.}

\item{...}{Optional arguments to \code{\link{interfaceplot}}.}
}
\value{
A list with interface tables for each module
}
\description{
Function that applies \code{\link{interfaceplot}} for a whole model and
all its modules.
}
\seealso{
\code{\link{codeCheck}},\code{\link{interfaceplot}}
}
\author{
Jan Philipp Dietrich
}
