% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path}
\alias{path}
\title{path}
\usage{
path(..., ftype = NULL)
}
\arguments{
\item{...}{the folders and the file name that should be pasted to a
file/folder path}

\item{ftype}{file type}
}
\value{
A string containing the path combined of folder, filename and
filetype
}
\description{
Small function to build a consistent path-string based on folder, filename
and filetype. The function makes sure that slashes and the dot for the file
ending are set correctly (you can supply your folder name either with or
without a tailing slash in it. It does not matter.
}
\author{
Jan Philipp Dietrich
}
