% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{post_process}
\alias{post_process}
\title{Generative Bootstrap Sampler for Logisti Regression Model}
\usage{
post_process(samples_GMS, alpha = 0.95, theta_hat = NULL,
thre = 0.005)
}
\arguments{
\item{samples_GMS}{an output object from "GBS_Sampling" function.}

\item{alpha}{confidence level.}

\item{theta_hat}{the point estimator of the full data.}

\item{thre}{threshold to sparcify LASSO solutions. This value will be used to implement "StabSel".}
}
\description{
Post-processing procedure to evaluate CIs and bias-corrected estimators.
}
\details{
#@importFrom Rcpp evalCpp
}
\examples{
#post_process(samples_GBS, theta_hat = theta_hat, thre=0.001)
}
\seealso{
\code{\link{GMS_Sampling}}, \code{\link{GMS}}
}
\author{
Minsuk Shin, Jun Liu and Shijie Wang
}
