% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDeclarations.R
\name{readDeclarations}
\alias{readDeclarations}
\title{readDeclarations}
\usage{
readDeclarations(
  file,
  unlist = TRUE,
  types = c("scalar", "(positive |)variable", "parameter", "table", "equation", "set")
)
}
\arguments{
\item{file}{A gams file or a vector containing GAMS code.}

\item{unlist}{A logical indicating whether the output should be returned as
a list separated by object type or a matrix.}

\item{types}{of declarations to be read.}
}
\value{
Either a list of declared objects or a matrix containing object
name, the sets the object depends on and the description.
}
\description{
Reads all declarations given in a GAMS code and returns them.
}
\seealso{
\code{\link{codeCheck}}
}
\author{
Jan Philipp Dietrich
}
