% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFileOrVector.R
\name{readFileOrVector}
\alias{readFileOrVector}
\title{readFileOrVector}
\usage{
readFileOrVector(file)
}
\arguments{
\item{file}{A file name or a character vector.}
}
\value{
A character vector of the inputs of the file.
}
\description{
Input: a file name or a character vector. Output: a character vector of the contents of the file, or
the character vector unchanged.
}
\author{
Mika Pflüger
}
