% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readParameterValues.R
\name{readParameterValues}
\alias{readParameterValues}
\title{readParameterValues}
\usage{
readParameterValues(code, parameters)
}
\arguments{
\item{code}{A vector with GAMS code.}

\item{parameters}{A vector with GAMS parameter names}
}
\value{
A vector of values the parameters are set to with parameter names as
names.
}
\description{
Read values of given parameters from a given GAMS code section.
}
\details{
The GAMS code section should contain statements of the form
parameter = value;
for all the given parameters.
}
\seealso{
\code{\link{readDeclarations}}
}
\author{
Mika Pflüger
}
