% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSetglobals.R
\name{readSetglobals}
\alias{readSetglobals}
\title{readSetglobals}
\usage{
readSetglobals(file)
}
\arguments{
\item{file}{A gams file or a vector containing GAMS code.}
}
\value{
A vector of values the setglobal variables are set to with setglobal
variables as names.
}
\description{
Reads all setglobals given in a GAMS code and returns them.
}
\seealso{
\code{\link{readDeclarations}}
}
\author{
Jan Philipp Dietrich
}
