% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSetItems.R
\name{readSetItems}
\alias{readSetItems}
\title{readSetItems}
\usage{
readSetItems(codeFile, warn = FALSE, name = NULL)
}
\arguments{
\item{codeFile}{a GAMS code file}

\item{warn}{A boolean indicating if warnings should be displayed when parsing
a sections fails}

\item{name}{A name indicating what collection of code files this is (e.g.
module name). Only needed for more expressive warnings.}
}
\description{
Parses the sets for a given code file and stores them in a named list.
}
\author{
Falk Benke
}
