% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSettings.R
\name{readSettings}
\alias{readSettings}
\title{readSettings}
\usage{
readSettings(file)
}
\arguments{
\item{file}{A gams file or a vector containing GAMS code.}
}
\value{
A vector of parameter values and their names.
}
\description{
Reads all settings (parameters, globals, and their respective values)
in the given GAMS code.
}
\seealso{
\code{\link{readDeclarations}}
}
\author{
Mika Pflüger
}
