% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveConfig.R
\name{saveConfig}
\alias{saveConfig}
\title{Save Config}
\usage{
saveConfig(cfg, file = NULL)
}
\arguments{
\item{cfg}{Input config which should be saved}

\item{file}{A character string naming a file. If set to NULL the YAML output
will be returned by the function.}
}
\description{
Saves config in YAML format so that it can be read in again via
\code{\link{loadConfig}}.
}
\examples{
cfg <- list(input = c(data1 = "blub.tgz", data2 = "bla.tgz"), mode = "default")
saveConfig(cfg)
}
\seealso{
\code{\link{loadConfig}}
}
\author{
Jan Philipp Dietrich
}
