% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectScript.R
\name{selectScript}
\alias{selectScript}
\title{selectScript}
\usage{
selectScript(folder = ".", ending = "R")
}
\arguments{
\item{folder}{Folder in which the files/scripts are located which should
be selected from.}

\item{ending}{File ending of the files to be selected (without dot)}
}
\value{
A vector of paths to files selected by the user
}
\description{
Functions which allows for interactive selection of scripts/files.
}
\author{
Jan Philipp Dietrich
}
