% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tardir.R
\name{tardir}
\alias{tardir}
\title{Create tgz archive from directory}
\usage{
tardir(dir = ".", tarfile = "data.tgz")
}
\arguments{
\item{dir}{directory from which the tar file should be generated}

\item{tarfile}{name of the archive the data should be written to (tgz file)}
}
\description{
Creates a tgz from all files in a directory
}
\examples{
# copy dummymodel to temporary directory and compress it
file.copy(system.file("dummymodel",package="gms"),tempdir(), recursive = TRUE)
model   <- paste0(tempdir(),"/dummymodel")
archive <- paste0(tempdir(),"/dummymodel.tgz")
tardir(model,archive)

}
\author{
Jan Philipp Dietrich
}
