% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_fulldataOutput.R
\name{update_fulldataOutput}
\alias{update_fulldataOutput}
\title{update_fulldataOutput}
\usage{
update_fulldataOutput(
  modelpath = ".",
  modulepath = "modules",
  corepath = "core",
  loopset = "t"
)
}
\arguments{
\item{modelpath}{Path of the Model version that should be updated (main
folder).}

\item{modulepath}{Module path within the model (relative to the model main
folder)}

\item{corepath}{Core path within the model (relative to the model main
folder)}

\item{loopset}{Set over which loop runs}
}
\description{
Creates GAMS code which stores automatically the levels and marginals of all
equations and variables in time depending parameters.
}
\seealso{
\code{\link{fulldataOutput}},\code{\link{replace_in_file}}
}
\author{
Jan Philipp Dietrich, Felicitas Beier
}
