% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateInterfaceMapping.R
\name{updateInterfaceMapping}
\alias{updateInterfaceMapping}
\title{updateInterfaceMapping}
\usage{
updateInterfaceMapping(path = ".", modulepath = "modules")
}
\arguments{
\item{path}{path of the main folder of the model}

\item{modulepath}{path to the module folder relative to "path"}
}
\description{
Function to update the mapping between interfaces and their origin modules.
}
\seealso{
\code{\link{codeCheck}}
}
\author{
Jan Philipp Dietrich
}
