% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfiledestinations.R
\name{getfiledestinations}
\alias{getfiledestinations}
\title{getfiledestinations}
\usage{
getfiledestinations(path = ".", ignoreFolders = "renv")
}
\arguments{
\item{path}{main model folder}

\item{ignoreFolders}{folders to be ignored by the function, additionally to
directories listed in \code{.gitignore} (by default only "renv").}
}
\description{
Create file2destination mapping based on information from the model, ignoring
top-level directories listed in \code{.gitignore}.
}
\author{
Jan Philipp Dietrich, David Klein
}
